tests/wpt/tests/css/css-flexbox/parsing/flex-direction-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing flex-direction with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-direction">
<meta name="assert" content="flex-direction supports only the grammar 'row | row-reverse | column | column-reverse'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flex-direction", "auto");
test_invalid_value("flex-direction", "column row-reverse");
</script>
</body>
</html>
